requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==1.4.49
python-jose[cryptography]==3.3.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.24.3
ultralytics==8.0.196
databases==0.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
aiofiles==23.2.1